refactor(vllm): move SourceIdentity construction into the vllm engine
